[Hotfix] default 이미지 다시 부여 세부사항 - default 이미지가 이미 DB에 저장되어 있으면 default 이미지는 삭제되지 않고 그냥 true 반환됨.

+++ src/main/java/org/nurim/nurim/service/MemberImageService.java
@@ -54,15 +54,24 @@ public class MemberImageService {
         if (memberImageOptional.isPresent()) {
             MemberImage memberImage = memberImageOptional.get();
 
+            // 이미지의 url과 key가 default와 일치하는지 확인
+            String imageUrl = memberImage.getMemberProfileImage();
+            String imageKey = memberImage.getProfileName();
+
+            // 기본 이미지로 변경할 이미지 경로
+            String defaultImageUrl = "https://nurimplus.s3.ap-northeast-2.amazonaws.com/images/b706c0f7-625a-485f-9d6e-2358822208bb.jpeg";
+            String defaultKey = "images/b706c0f7-625a-485f-9d6e-2358822208bb.jpeg";
+
+            if (imageUrl.equals(defaultImageUrl) && imageKey.equals(defaultKey)) {
+                // 이미지의 url과 key가 default와 일치하면 삭제 로직을 적용하지 않음
+                return true;
+            }
+
             // S3에서 파일 삭제
             boolean isRemovedFromS3 = fileUploadService.deleteFile(memberImage.getProfileName());
 
             // S3에서 이미지 삭제에 성공하면 DB에서 기본 이미지로 변경
             if (isRemovedFromS3) {
-                // 기본 이미지로 변경할 이미지 경로
-                String defaultImageUrl = "https://nurimplus.s3.ap-northeast-2.amazonaws.com/images/b706c0f7-625a-485f-9d6e-2358822208bb.jpeg";
-                String defaultKey = "images/b706c0f7-625a-485f-9d6e-2358822208bb.jpeg";
-
                 // DB에서 해당 회원의 이미지를 기본 이미지로 변경
                 memberImage.setMemberProfileImage(defaultImageUrl);
                 memberImage.setProfileName(defaultKey);
